+++ test/spec/validator/between.js
@@ -56,6 +56,12 @@ describe('between', function() {
         $('#betweenForm').bootstrapValidator('destroy').remove();
     });
 
+    it('not a number', function() {
+        this.$age.val('50abc');
+        this.bv.validate();
+        expect(this.bv.isValid()).toEqual(false);
+    });
+
     it('compare to value', function() {
         this.$age.val(10);
         this.bv.validate();