+++ test/spec/validator/greaterThan.js
@@ -39,6 +39,12 @@ describe('greaterThan', function() {
         $('#greaterThanForm').bootstrapValidator('destroy').remove();
     });
 
+    it('not a number', function() {
+        this.$age.val('20abc');
+        this.bv.validate();
+        expect(this.bv.isValid()).toEqual(false);
+    });
+
     it('compare to value', function() {
         this.$age.val(10);
         this.bv.validate();